Parcially fixes #17
This partially fixes #17 . Since we don't have the ID of the user resetting their passwords, old passwords cannot be checked. For this to work, the PasswordModule needs to be modified to allow the member to be accessible. Or maybe the member can be retrieved via the reset token, but that would mean doing the same things as with the password module again.